Move Listener in UiBundle as it can concern both Admin and Shop routes
