--- src/Sylius/Bundle/ShopBundle/EventListener/AlreadyLoggedInUserRegistrationListener.php
@@ -1,42 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Sylius Sp. z o.o.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Sylius\Bundle\ShopBundle\EventListener;
-
-use Sylius\Bundle\ResourceBundle\Event\ResourceControllerEvent;
-use Symfony\Component\HttpFoundation\RedirectResponse;
-use Symfony\Component\HttpFoundation\RequestStack;
-use Symfony\Component\Routing\RouterInterface;
-use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
-
-final class AlreadyLoggedInUserRegistrationListener
-{
-    public function __construct(
-        private AuthorizationCheckerInterface $authorizationChecker,
-        private RequestStack $requestStack,
-        private RouterInterface $router
-    ) {
-    }
-
-    public function handleAlreadyConnectedUser(ResourceControllerEvent $event): void
-    {
-        if (!$this->authorizationChecker->isGranted('IS_AUTHENTICATED_REMEMBERED')) {
-            return;
-        }
-
-        $alreadyLoggedInRedirectRoute = $this->requestStack->getMainRequest()->attributes->get('_sylius')['logged_in_route'] ?? null;
-        if (null !== $alreadyLoggedInRedirectRoute) {
-            $event->setResponse(new RedirectResponse($this->router->generate($alreadyLoggedInRedirectRoute)));
-        }
-    }
-}
