+++ src/Sylius/Bundle/UiBundle/Resources/config/services.xml
@@ -16,6 +16,7 @@
         <import resource="services/context_provider.xml" />
         <import resource="services/controller.xml" />
         <import resource="services/form.xml" />
+        <import resource="services/listeners.xml" />
         <import resource="services/template_event.xml" />
         <import resource="services/twig.xml" />
     </imports>
