+++ src/Sylius/Bundle/UiBundle/Resources/config/services/listeners.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Sylius Sp. z o.o.
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+    <services>
+        <defaults public="true" />
+        <service id="sylius.listener.already_logged_in_user_redirection" class="Sylius\Bundle\UiBundle\EventListener\AlreadyLoggedInUserRedirectionListener">
+            <argument type="service" id="security.authorization_checker"/>
+            <argument type="service" id="request_stack" />
+            <argument type="service" id="router" />
+            <tag name="kernel.event_listener" event="kernel.request" method="handleAlreadyConnectedUser" />
+        </service>
+    </services>
+</container>
